Remove seq2seq_autoencoder.py to reduce complexity.
